private: Avoid unnecessary conversions in shadow computation
The type was changed from double to float for no apparent reason
when moving code in commit 5060081db5f.

https://bugzilla.gnome.org/show_bug.cgi?id=788627

+++ src/st/st-private.c
@@ -247,7 +247,7 @@ blur_pixels (guchar  *pixels_in,
              gint    *rowstride_out)
 {
   guchar *pixels_out;
-  float   sigma;
+  gdouble sigma;
 
   /* The CSS specification defines (or will define) the blur radius as twice
    * the Gaussian standard deviation. See:
